software/*/mmcMailbox: fix MMC init counter comparison

+++ software/src/mmcMailbox.h
@@ -13,8 +13,8 @@ int mmcMailboxRead(unsigned int address);
 uint32_t *mmcMailboxFetchSysmon(uint32_t *ap);
 int getU28temperature(void);
 int getU29temperature(void);
-int getMMCPG3Count(void);
-int getMMCPG4Count(void);
+uint16_t getMMCPG3Count(void);
+uint16_t getMMCPG4Count(void);
 int mmcMailboxIsInit(void);
 
 #endif /* _MMC_MAILBOX_H_ */
